Updated install script and readme.

+++ install.sh
@@ -38,30 +38,82 @@ function install_linux {
 	function install_package {
 		PACKAGE_MANAGER=`which apt-get`
 		if [[ ! -z "${PACKAGE_MANAGER}" ]]; then
-			${PACKAGE_MANAGER} install ${1} ||
+			${PACKAGE_MANAGER} --yes --force-yes install ${@} ||
 				error "could not install ${1}!"
 		fi
 	}
 	header "Installing for Linux"
-	exit 1
-	wget "http://download.teamviewer.com/download/teamviewer_i386.deb"
-	dpkg -i -y "teamviewer_i386.deb"
-	rm -y "teamviewer_i386.deb"
-	if [[ -z "$( which kivy )" ]]; then
-		install_package -y build-essential
-		install_package -y python-pip
-		install_package -y python-dev
-		install_package -y git
-		install_package -y u-boot-tools
 
+	if [[ -z "$( which teamviewer)" ]]; then
+		install_package libc6:i386 \
+						libgcc1:i386 \
+						libasound2:i386 \
+						libexpat1:i386 \
+						libfontconfig1:i386 \
+						libfreetype6:i386 \
+						libjpeg62:i386 \
+						libpng12-0:i386 \
+						libsm6:i386 \
+						libxdamage1:i386 \
+						libxext6:i386 \
+						libxfixes3:i386 \
+						libxinerama1:i386 \
+						libxrandr2:i386 \
+						libxrender1:i386 \
+						libxtst6:i386 \
+						zlib1g:i386
+		wget "http://download.teamviewer.com/download/teamviewer_i386.deb"
+		dpkg -i "teamviewer_i386.deb"
+		rm -y "teamviewer_i386.deb"
+	fi
+	if [[ -z "$( which pip )" ]]; then
+		install_package python-pip \
+						python-dev
+	fi
+	if [[ -z "$( which gcc )" ]]; then
+		install_package build-essential \
+						git
+	fi
+	if [[ -z "$( which mkimage )" ]]; then
+		install_package u-boot-tools
+	fi
+	if [[ -z "$( which kivy )" ]]; then
+		install_package mesa-common-dev \
+						libgl1-mesa-dev \
+						python-setuptools \
+						python-pygame \
+						python-opengl \
+						python-gst0.10 \
+						python-enchant \
+						gstreamer0.10-plugins-good \
+						libgles2-mesa-dev
 		PIP=`which pip`
-		${PIP} install cython || error "could not install cython!"
+		${PIP} install --upgrade Cython==0.21 || error "could not install cython!"
 		${PIP} install kivy || error "could not install kivy!"
 		${PIP} install libusb1 || error "could not install libusb1!"
+		sudo ln -s /usr/bin/python2.7 /usr/bin/kivy
 	fi
+}
+function install_flasher {
+	if [[ ! -d "flasher" ]];then
+		git clone https://github.com/NextThingCo/CHIP-flasher.git flasher
+	fi
+	if [[ ! -d "flasher/tools" ]];then
+		git clone https://github.com/NextThingCo/CHIP-tools flasher/tools
+	fi
+	if [[ ! -f "flasher/sunxi-tools/fel" ]];then
+		install_package libusb-1.0-0-dev android-tools-fastboot
+		if [[ ! -d "flasher/sunxi-tools" ]];then
+			git clone https://github.com/linux-sunxi/sunxi-tools flasher/sunxi-tools
+		fi
+		make -C flasher/sunxi-tools fel
+		ln -s "$(pwd)/flasher/sunxi-tools/fel" /usr/bin/fel
+	fi
+	chmod -R 777 flasher
+	DISPLAY=:0 kivy flasher/main.py
 }
 
 case "${OS}" in
-	Darwin) install_darwin ;;
-	Linux) install_linux ;;
+	Darwin) install_darwin; install_flasher ;;
+	Linux) install_linux; install_flasher ;;
 esac
